CreateMenu logic moved in WorkFooter
